refactor: refactor styles for components
