+++ src/components/VoteCard/VoteCard.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from "react";
-import styles from '@/styles/cards.module.scss';
+import styles from './VoteCard.module.scss';
 import { CardType } from "@/untils/types";
 
 type Props = {
